Translate Bugzilla XML into GitHub JSON.
This translates into GitHub's *export* format, which unfortunately
is not the same as (or even a superset of) GitHub's *import* format.
But this is good because it means we can compare the
exported-from-Bugzilla-and-transformed JSON files pretty much directly
against the exported-from-GitHub JSON files, to verify that the import
(when it is implemented) has gone right.